# docs/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: WeddingOS Unified API
  version: 1.0.0
  description: |
    Unified OpenAPI 3.0 specification for the WeddingOS microservices platform.

    This document was assembled from the actual service route files and related
    Zod / Prisma / Pydantic definitions in the monorepo. Paths are documented as
    gateway-facing routes using the service prefixes requested by the user
    (`/auth`, `/users`, `/vendors`, `/bookings`, `/payments`, `/execution`,
    `/notifications`, `/reviews`, `/search`, `/media`, `/chat`, `/ai`).

    Notes:
    - Most services return a common JSON envelope with `success`, `data`, and often `meta`.
    - Internal routes may accept either a bearer token with admin role or `x-internal-api-key`.
    - Chat real-time Socket.IO events are described in tag descriptions and README, while this spec focuses on HTTP endpoints.
servers:
  - url: http://localhost:8000
    description: Local API gateway
  - url: https://api-staging.weddingos.in
    description: Staging
  - url: https://api.weddingos.in
    description: Production
security:
  - bearerAuth: []
...
paths:
...
  /vendors/admin/list:
    get:
      tags: [Vendors]
      summary: Admin list vendors
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit100'
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: category
          schema: { type: string }
        - in: query
          name: q
          schema: { type: string }
      responses:
        '200':
          description: Paginated vendor admin list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminVendorListEnvelope'
...
  /bookings/admin/list:
    get:
      tags: [Bookings]
      summary: Admin list bookings
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit100'
        - in: query
          name: status
          schema:
            $ref: '#/components/schemas/BookingStatus'
      responses:
        '200':
          description: Paginated bookings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminBookingListEnvelope'
...
  /payments/admin/list:
    get:
      tags: [Payments]
      summary: Admin list payments
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit100'
        - in: query
          name: status
          schema:
            $ref: '#/components/schemas/PaymentStatus'
      responses:
        '200':
          description: Paginated payment list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminPaymentListEnvelope'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    internalApiKey:
      type: apiKey
      in: header
      name: x-internal-api-key
  parameters:
    Page:
      in: query
      name: page
      schema:
        type: integer
        minimum: 1
        default: 1
    Limit100:
      in: query
      name: limit
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
  schemas:
    ApiMeta:
      type: object
      properties:
        requestId:
          type: string
          nullable: true
        timestamp:
          type: string
          format: date-time
...
    VendorCategory:
      type: string
      enum:
        - PHOTOGRAPHER
        - VIDEOGRAPHER
        - CATERER
        - DECORATOR
        - VENUE
        - DJ_SOUND
        - BAND_ENTERTAINMENT
        - BRIDAL_MAKEUP
        - GROOM_MAKEUP
        - MEHENDI
        - PANDIT_PRIEST
        - WEDDING_PLANNER
        - INVITATIONS
        - CHOREOGRAPHER
        - BARTENDER
        - LIGHTING
        - FIREWORKS
        - TENT_HOUSE
        - TRANSPORTATION
        - FLORIST
    VendorStatus:
      type: string
      enum: [DRAFT, PENDING_REVIEW, ACTIVE, SUSPENDED, BLACKLISTED, REJECTED, PENDING_KYC]
    PackageType:
      type: string
      enum: [BASIC, STANDARD, PREMIUM, CUSTOM]
    BookingStatus:
      type: string
      enum: [ENQUIRY, QUOTE_SENT, QUOTE_ACCEPTED, ADVANCE_PENDING, ADVANCE_PAID, CONFIRMED, CHECKIN, COMPLETED, CANCELLED_BY_CUSTOMER, CANCELLED_BY_VENDOR, DISPUTED, REFUNDED]
    EventType:
      type: string
      enum: [WEDDING_CEREMONY, RECEPTION, ENGAGEMENT, HALDI, MEHNDI, SANGEET, BACHELOR_PARTY, PRE_WEDDING_SHOOT, BIRTHDAY, ANNIVERSARY, CORPORATE, OTHER]
    PaymentStatus:
      type: string
      enum: [CREATED, PENDING, CAPTURED, FAILED, REFUNDED, PARTIALLY_REFUNDED]
...
    VendorPackage:
      type: object
      properties:
        id: { type: string, format: uuid }
        vendorId: { type: string, format: uuid }
        packageType:
          $ref: '#/components/schemas/PackageType'
        name: { type: string }
        description: { type: string, nullable: true }
        priceFromPaise: { type: integer }
        priceUpToPaise: { type: integer, nullable: true }
        isCustomQuote: { type: boolean }
        inclusions:
          type: array
          items: { type: string }
        exclusions:
          type: array
          items: { type: string }
        deliverables:
          type: array
          items: { type: string }
        isActive: { type: boolean }
        sortOrder: { type: integer }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    PortfolioItem:
      type: object
      properties:
        id: { type: string, format: uuid }
        vendorId: { type: string, format: uuid }
        s3Key: { type: string }
        publicUrl: { type: string }
        thumbUrl: { type: string, nullable: true }
        caption: { type: string, nullable: true }
        mediaType: { type: string }
        sortOrder: { type: integer }
        createdAt: { type: string, format: date-time }
    AvailabilityBlock:
      type: object
      properties:
        id: { type: string, format: uuid }
        vendorId: { type: string, format: uuid }
        blockedDate: { type: string, format: date-time }
        reason: { type: string, nullable: true }
    Vendor:
      type: object
      properties:
        id: { type: string, format: uuid }
        userId: { type: string, format: uuid }
        businessName: { type: string }
        slug: { type: string }
        category:
          $ref: '#/components/schemas/VendorCategory'
        subCategories:
          type: array
          items:
            $ref: '#/components/schemas/VendorCategory'
        status:
          $ref: '#/components/schemas/VendorStatus'
        city: { type: string }
        state: { type: string }
        pincode: { type: string }
        serviceCities:
          type: array
          items: { type: string }
        tagline: { type: string, nullable: true }
        description: { type: string, nullable: true }
        yearsExperience: { type: integer, nullable: true }
        teamSize: { type: integer, nullable: true }
        coverPhoto: { type: string, nullable: true }
        logoUrl: { type: string, nullable: true }
        whatsappNumber: { type: string, nullable: true }
        websiteUrl: { type: string, nullable: true }
        instagramUrl: { type: string, nullable: true }
        gstNumber: { type: string, nullable: true }
        panNumber: { type: string, nullable: true }
        bankAccountNo: { type: string, nullable: true }
        bankIfsc: { type: string, nullable: true }
        bankAccountName: { type: string, nullable: true }
        avgRating: { type: number, format: float }
        reviewCount: { type: integer }
        bookingCount: { type: integer }
        responseRatePercent: { type: integer }
        avgResponseHours: { type: number, format: float }
        plusMember: { type: boolean }
        isFeatured: { type: boolean }
        adminNote: { type: string, nullable: true }
        packages:
          type: array
          items:
            $ref: '#/components/schemas/VendorPackage'
        portfolio:
          type: array
          items:
            $ref: '#/components/schemas/PortfolioItem'
        availability:
          type: array
          items:
            $ref: '#/components/schemas/AvailabilityBlock'
        tags:
          type: array
          items:
            type: object
            properties:
              tag: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    AdminVendorListEnvelope:
      type: object
      properties:
        success: { type: boolean, example: true }
        data:
          type: object
          properties:
            vendors:
              type: array
              items:
                $ref: '#/components/schemas/Vendor'
        meta:
          allOf:
            - $ref: '#/components/schemas/ApiMeta'
            - type: object
              properties:
                total: { type: integer }
                page: { type: integer }
                limit: { type: integer }
                pages: { type: integer }
...
    Booking:
      type: object
      properties:
        id: { type: string, format: uuid }
        bookingNumber: { type: string, nullable: true }
        customerId: { type: string, format: uuid }
        vendorId: { type: string, format: uuid }
        packageId: { type: string, format: uuid, nullable: true }
        status:
          $ref: '#/components/schemas/BookingStatus'
        eventDate: { type: string, format: date-time }
        eventType:
          $ref: '#/components/schemas/EventType'
        eventCity: { type: string }
        requirements: { type: string, nullable: true }
        guestCount: { type: integer, nullable: true }
        specialNotes: { type: string, nullable: true }
        quotedAmountPaise: { type: integer, nullable: true }
        finalAmountPaise: { type: integer, nullable: true }
        vendorQuoteNote: { type: string, nullable: true }
        paymentId: { type: string, format: uuid, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    AdminBookingListEnvelope:
      type: object
      properties:
        success: { type: boolean, example: true }
        data:
          type: object
          properties:
            bookings:
              type: array
              items:
                $ref: '#/components/schemas/Booking'
        meta:
          allOf:
            - $ref: '#/components/schemas/ApiMeta'
            - type: object
              properties:
                total: { type: integer }
                page: { type: integer }
                limit: { type: integer }
                pages: { type: integer }
...
    Payment:
      type: object
      properties:
        id: { type: string, format: uuid }
        bookingId: { type: string, format: uuid }
        customerId: { type: string, format: uuid }
        vendorId: { type: string, format: uuid }
        amountPaise: { type: integer }
        currency: { type: string }
        description: { type: string, nullable: true }
        razorpayOrderId: { type: string }
        razorpayPaymentId: { type: string, nullable: true }
        status:
          $ref: '#/components/schemas/PaymentStatus'
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    AdminPaymentListEnvelope:
      type: object
      properties:
        success: { type: boolean, example: true }
        data:
          type: object
          properties:
            payments:
              type: array
              items:
                $ref: '#/components/schemas/Payment'
        meta:
          allOf:
            - $ref: '#/components/schemas/ApiMeta'
            - type: object
              properties:
                total: { type: integer }
                page: { type: integer }
                limit: { type: integer }
                pages: { type: integer }
...